+++ src/datatypes/function.c
@@ -1,6 +1,6 @@
 /*
 function.c - Une
-Modified 2021-08-13
+Modified 2021-09-28
 */
 
 /* Header-specific includes. */
@@ -34,29 +34,29 @@ Call result.
 une_result une_datatype_function_call(une_error *error, une_interpreter_state *is, une_node *call, une_result function, une_result args)
 {
   /* Get function. */
-  une_function *fn = (une_function*)function.value._vp;
-  assert(fn != NULL);
+  size_t fn = (size_t)function.value._int;
+  assert(fn <= is->functions_size);
   
   /* Ensure number of arguments matches number of required parameters. */
   UNE_UNPACK_RESULT_LIST(args, args_p, args_count);
-  if (fn->params_count != args_count) {
+  if ((is->functions)[fn].params_count != args_count) {
     *error = UNE_ERROR_SET(UNE_ET_CALLABLE_ARG_COUNT, call->pos);
     return une_result_create(UNE_RT_ERROR);
   }
   
   /* Create function context. */
   une_context *parent = is->context;
-  is->context = une_context_create(fn, UNE_SIZE_VARIABLE_BUF, UNE_SIZE_FUNCTION_BUF);
+  is->context = une_context_create((ptrdiff_t)fn, UNE_SIZE_VARIABLE_BUF);
   is->context->parent = parent;
 
   /* Define parameters. */
-  for (size_t i=0; i<fn->params_count; i++) {
-    une_variable *var = une_variable_create(is->context, fn->params[i]);
+  for (size_t i=0; i<(is->functions)[fn].params_count; i++) {
+    une_variable *var = une_variable_create(is->context, ((is->functions)[fn].params)[i]);
     var->content = une_result_copy(args_p[i+1]);
   }
 
   /* Interpret body. */
-  une_result result = une_interpret(error, is, fn->body);
+  une_result result = une_interpret(error, is, (is->functions)[fn].body);
   is->should_return = false;
 
   /* Return to parent context. */