+++ src/types/context.c
@@ -1,6 +1,6 @@
 /*
 context.c - Une
-Modified 2021-08-14
+Modified 2021-09-28
 */
 
 /* Header-specific includes. */
@@ -12,7 +12,7 @@ Modified 2021-08-14
 /*
 Allocates, initializes, and returns a pointer to a une_context struct.
 */
-une_context *une_context_create(une_function *function, size_t variables_size, size_t functions_size)
+une_context *une_context_create(ptrdiff_t function, size_t variables_size)
 {
   /* Allocate une_context. */
   une_context *context = malloc(sizeof(*context));
@@ -24,13 +24,9 @@ une_context *une_context_create(une_function *function, size_t variables_size, s
     .function = function,
     .variables = malloc(variables_size*sizeof(*context->variables)),
     .variables_size = variables_size,
-    .variables_count = 0,
-    .functions = malloc(functions_size*sizeof(*context->functions)),
-    .functions_size = functions_size,
-    .functions_count = 0,
+    .variables_count = 0
   };
   verify(context->variables);
-  verify(context->functions);
   
   return context;
 }
@@ -60,13 +56,6 @@ void une_context_free(une_context *context)
     free(context->variables);
   }
   
-  /* Free une_function buffer. */
-  if (context->functions != NULL) {
-    for (size_t i=0; i<context->functions_count; i++)
-      une_function_free((context->functions)[i]);
-    free(context->functions);
-  }
-
   free(context);
 }
 
@@ -153,32 +142,4 @@ __une_variable_itf(une_variable_find_or_create_global)
   
   /* une_variable doesn't exist yet, create it. */
   return une_variable_create(context, name);
-}
-
-/*
-Initializes a new une_function in a une_context's function buffer.
-*/
-une_function *une_function_create(une_context *context, char *definition_file, une_position definition_point)
-{
-  /* Ensure sufficient space in une_function buffer. */
-  if (context->functions_count >= context->functions_size) {
-    context->functions_size *= 2;
-    context->functions = realloc(context->functions, context->functions_size*sizeof(*context->functions));
-    verify(context->functions);
-  }
-  
-  /* Initialize une_function. */
-  une_function *fn = malloc(sizeof(*fn));
-  verify(fn);
-  (context->functions)[(context->functions_count)++] = fn;
-  *fn = (une_function){
-    .definition_file = strdup(definition_file),
-    .definition_point = definition_point,
-    .params_count = 0,
-    .params = NULL,
-    .body = NULL
-  };
-  verify(fn->definition_file);
-  
-  return fn;
 }