+++ src/types/context.h
@@ -1,6 +1,6 @@
 /*
 context.h - Une
-Modified 2021-09-14
+Modified 2021-09-28
 */
 
 #ifndef UNE_CONTEXT_H
@@ -15,13 +15,10 @@ Holds information that changes depending on the execution context.
 */
 typedef struct _une_context {
   struct _une_context *parent;
-  une_function *function;
+  ptrdiff_t function;
   size_t variables_size;
   size_t variables_count;
   une_variable *variables;
-  size_t functions_size;
-  size_t functions_count;
-  une_function **functions;
 } une_context;
 
 /*
@@ -33,7 +30,7 @@ Variable interface function template.
 */
 #define __une_variable_itf(__id) une_variable *(__id)(une_context *context, wchar_t *name) 
 
-une_context *une_context_create(une_function *function, size_t variables_size, size_t functions_size);
+une_context *une_context_create(ptrdiff_t function, size_t variables_size);
 void une_context_free_children(une_context *parent, une_context *youngest_child);
 void une_context_free(une_context *context);
 
@@ -42,7 +39,5 @@ __une_variable_itf(une_variable_find);
 __une_variable_itf(une_variable_find_global);
 __une_variable_itf(une_variable_find_or_create);
 __une_variable_itf(une_variable_find_or_create_global);
-
-une_function *une_function_create(une_context *context, char *definition_file, une_position definition_point);
 
 #endif /* !UNE_CONTEXT_H */